[#131,#132] IRODSFilesystem: Fix retrieval of data object permissions
This commit fixes a bug in which empty strings are returned by GenQuery2
when the user does not have permissions on a data object. This situation
is specific to iRODS 5.0.2 (https://github.com/irods/irods/issues/8546).

This commit fixes the issue by filtering out the rows containing empty
strings from the results.

To avoid truncated results, the LIMIT keyword is explicitly set to
10000.